requirements.txt
aiohttp==3.8.0
fastapi==0.70.0
pydantic==1.8.2
requests==2.26.0
starlette==0.16.0
uvicorn==0.15.0
python-dotenv==0.19.2
pytest==7.1.1